change userId type from nullable Int to Long

+++ app/src/main/java/com/aashish/bookshelf/repository/AuthManager.kt
@@ -1,8 +1,8 @@
 package com.aashish.bookshelf.repository
 
 interface AuthManager {
-    fun getLastLoginUserId(): Int
-    fun updateLastLoginUserId(userId: Int)
+    fun getLastLoginUserId(): Long
+    fun updateLastLoginUserId(userId: Long)
 
     fun resetLastLoginUserId()
 }
